Refactored code (split router and controller logic. Added and started work on validatePage.js, which contains the core logic of WACG analysis
