+++ backend/routes/Auth.routes.js
@@ -1,60 +1,9 @@
-import express from 'express';
-import bcryptjs from 'bcryptjs';
-import jsonwebtoken from 'jsonwebtoken';
-import prisma from '../database/init.js';
+import express from "express";
+import { register, login } from "../controllers/Auth.controller.js";
 
-const bcrypt = bcryptjs;
-const jwt = jsonwebtoken;
 const router = express.Router();
 
-// Register endpoint
-router.post('/register', async (req, res) => {
-    const { email, password } = req.body;
-    try {
-        if (!email || !password) {
-            return res.status(400).json({ message: 'Email and password required.' });
-        }
-
-        const existingUser = await prisma.user.findUnique({ where: { email } });
-        if (existingUser) {
-            return res.status(409).json({ message: 'User already exists.' });
-        }
-
-        const hashedPassword = await bcrypt.hash(password, 12);
-        await prisma.user.create({
-            data: { email, password: hashedPassword, createdAt: new Date() }
-        });
-
-        res.status(201).json({ message: 'User registered successfully.' });
-    } catch (err) {
-        res.status(500).json({ message: err.message || 'Internal server error.' });
-    }
-});
-
-// Login endpoint
-router.post('/login', async (req, res) => {
-    const { email, password } = req.body;
-    try {
-        if (!email || !password)
-            return res.status(400).json({ message: 'Email and password required.' });
-
-        const user = await prisma.user.findUnique({ where: { email } });
-        if (!user)
-            return res.status(401).json({ message: 'Invalid credentials.' });
-
-        const isMatch = await bcrypt.compare(password, user.password);
-        if (!isMatch)
-            return res.status(401).json({ message: 'Invalid credentials.' });
-
-        const token = jwt.sign(
-            { email: user.email, id: user.id },
-            process.env.JWT_SECRET || 'secret',
-            { expiresIn: '1h' }
-        );
-        res.json({ token });
-    } catch (err) {
-        res.status(500).json({ message: 'Internal server error.' });
-    }
-});
+router.post("/register", register);
+router.post("/login", login);
 
 export default router;
